+++ ChangeLog.md
@@ -1,5 +1,8 @@
 # libmtev changes
 
+ * Added MTEV_MAYBE_REALLOC_WITH_TYPE for C++ compatibility.  The older
+   version is maintained for backwards compatibility.
+
 # 2
 
 ## 2.1